// controller/server/generated/control-protocol/tasklattice/guard/control/v1/ArtifactDependency.ts
// Original file: artifact.proto


/**
 * Version-pinned dependency required to load an artifact.
 */
export interface ArtifactDependency {
  'kind'?: (string);
  'name'?: (string);
  'version'?: (string);
}

/**
 * Version-pinned dependency required to load an artifact.
 */
export interface ArtifactDependency__Output {
  'kind': (string);
  'name': (string);
  'version': (string);
}